--channels: Allow dash in name when renaming channels.
Arguments to the "--channels" option are checked for the range
separator (a dash) first before handling a possible rename.

This makes it impossible to rename a channel to anything that
includes a dash.

So instead check for a rename (equals character) first and handle
a possible range later.

It makes no sense to rename a range of channels so report an error
if there is a dash character in the part preceding the equals sign,
except if a channel with this exact name exists.