--channels: Allow common prefix in channel ranges.
Selecting a range of channels only works if the channel names are
purely numeric, but not if they have a common, non-numeric
prefix (eg as in "D0", "D1", ..).

This patch allows to use such channels in ranges by checking for
a common prefix. The syntax to for example select channels "D2"
through to "D5" is "D2-5".

Also adjust the manpage to include the new syntax and fix the
example that actually didn't work as the fx2lafw driver uses
"D0", "D1", ... as channel names.